+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 * Property to determine the distance threshold between location updates
+### Changed
+* Adding a marker not longer updates all other markers in this location scene
 
 ### Changed
 * Renamed onRender listener -> onUpdate